test: Add comprehensive tests for both macros
This commit adds a significant number of new tests to improve the test coverage of the crate.

- Adds comprehensive tests for the new `#[default_args_attribute]` macro, covering complex cases like `async`, `unsafe`, `const`, generics, and `extern "C"` functions.
- Adds tests for more complex default value expressions, such as function calls, constants, and struct initializers, for both the attribute and function-like macros.
- Re-adds the `compile_fail` tests that were previously removed. These tests are now included as doc tests in `src/lib.rs` and verify that the macros correctly produce compile-time errors for invalid usage.
